разбиение родительского и дочерних классов
