+++ models/ActionCancel.php
@@ -1,0 +1,20 @@
+<?php
+namespace app\models;
+
+class ActionCancel extends Action
+{
+    public function getActionName()
+    {
+        return "Отменить";
+    }
+
+    public function getInnerActionName()
+    {
+        return "cancel";
+    }
+
+    public function checkUserRole($userID, $customerID, $executorID)
+    {
+        return ($userID == $customerID) ? true : false;
+    }
+}
