+++ models/ActionComplete.php
@@ -1,0 +1,20 @@
+<?php
+namespace app\models;
+
+class ActionComplete extends Action
+{
+    public function getActionName()
+    {
+        return "Выполнено";
+    }
+
+    public function getInnerActionName()
+    {
+        return "complete";
+    }
+
+    public function checkUserRole($userID, $customerID, $executorID)
+    {
+        return ($userID == $customerID) ? true : false;
+    }
+}
