+++ models/ActionRespond.php
@@ -1,0 +1,20 @@
+<?php
+namespace app\models;
+
+class ActionRespond extends Action
+{
+    public function getActionName()
+    {
+        return "Откликнуться";
+    }
+
+    public function getInnerActionName()
+    {
+        return "respond";
+    }
+
+    public function checkUserRole($userID, $customerID, $executorID)
+    {
+        return ($userID == $executorID) ? true : false;
+    }
+}
